Modified account controller to handle OpenID Attribute Exchange (AX)
This is more meant to be an example as to what needs to be fixed in
Redmine for Google/Yahoo/etc AX requests as SReg requests have been
depricated.  It's not the cleanest, but it's definitely functional.

Auto-registration works with Google and Yahoo for sure.  Registration
for other providers hasn't been tested.  After registration, it should
continue working as normal.

If using AX, usernames are generated by grabbing the first half of a
user's e-mail address.