Removed unintuitive logout from login method and added a redirect if is user already logged in.
The reason for this is that if you accidently go to the login page
and are logged in, you expect to be redirected as though you just
logged in, not logged off and required login again.

If the user is logged in already, it redirects to the home page.
If not, it displays the login page as intended.